Windows Powershell script to update BuildTools and Spigot.
Requires Git-Bash. https://git-scm.com/
It is recommended to create an 'Updater' folder in the Bitnami folder (C:\Bitnami\Updater) and place this script in that folder.
The script will download the latest version BuildTools and run it to download the latest version of Spigot. After Spigot.jar is built, the script will copy it to the Multicraft Daemon jar folder where it will be ready for the next server restart. Lastly, the script will remove all build file in order to free drive space.